Added docs

+++ source/ultrasonic.h
@@ -6,8 +6,11 @@
 #define TRIGGER_PIN 7 /* Used to trigger a measurement */
 #define ECHO_PIN 8 /* Output response of sensor */
 
+/// @brief Initializes the ultrasonic sensor on pins 7 and 8
 void initializeUltrasonic();
 
-float getUltrasonicOutput();
+/// @brief Returns the distance seen by the ultrasonic sensor in centimeters.
+/// @return The centimeters to the target.
+long getUltrasonicOutput();
 
 #endif
